Refactor config loading logic.
Functionality should be mostly the same. The only change i am aware of
is that empty values are now permitted, instead of disallowed. This will
cause a parse error when trying to parse as anything but strings, but
this is the same thing that would happen if you tried parsing "foo" as
an uint.